Split feature flag test module
