+++ integration/test/Test/FeatureFlags/GuestLinks.hs
@@ -1,0 +1,14 @@
+module Test.FeatureFlags.GuestLinks where
+
+import Test.FeatureFlags.Util
+import Testlib.Prelude
+
+testConversationGuestLinks :: (HasCallStack) => APIAccess -> App ()
+testConversationGuestLinks access =
+  mkFeatureTests "conversationGuestLinks" enabled
+    & addUpdate disabled
+    & addUpdate enabled
+    & runFeatureTests OwnDomain access
+
+testPatchGuestLinks :: (HasCallStack) => App ()
+testPatchGuestLinks = checkPatch OwnDomain "conversationGuestLinks" True enabled disabled
